+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## New Features
 
 * Improve synchronization with Toodledo
+* Trial license
+
+## Bug Fixes
+
+* Fix help button
+* Fix default column order
 
 # 0.6.0-beta